teamleader_issue update

+++ src/main/java/ai/group2/project_management_system/repository/UserRepository.java
@@ -10,6 +10,8 @@ import java.util.Optional;
 import java.util.Set;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
+
+
 }
